feat(revit): adds project info filter wip, and other minor improvements
