+++ DesktopUI/DesktopUI/Streams/Dialogs/FilterViews/ParameterFilterView.xaml
@@ -1,36 +1,40 @@
-﻿<UserControl x:Class="Speckle.DesktopUI.Streams.Dialogs.FilterViews.ParameterFilterView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:Speckle.DesktopUI.Streams.Dialogs.FilterViews"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:md="http://materialdesigninxaml.net/winfx/xaml/themes"
-             xmlns:s="https://github.com/canton7/Stylet"
-             d:DesignHeight="300"
-             d:DesignWidth="300"
-             mc:Ignorable="d">
+﻿<UserControl
+  x:Class="Speckle.DesktopUI.Streams.Dialogs.FilterViews.ParameterFilterView"
+  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+  xmlns:local="clr-namespace:Speckle.DesktopUI.Streams.Dialogs.FilterViews"
+  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+  xmlns:md="http://materialdesigninxaml.net/winfx/xaml/themes"
+  xmlns:s="https://github.com/canton7/Stylet"
+  d:DesignHeight="300"
+  d:DesignWidth="300"
+  mc:Ignorable="d">
   <Grid>
     <StackPanel>
-      <TextBlock Text="Add objects to this stream by parameter" />
-      <ComboBox Width="Auto"
-                Margin="0,16"
-                md:HintAssist.Hint="Parameter"
-                IsEditable="True"
-                IsEnabled="True"
-                ItemsSource="{Binding Filter.Values}"
-                SelectedItem="{Binding Filter.PropertyName}"
-                StaysOpenOnEdit="True" />
-      <ComboBox Width="Auto"
-                Margin="0,16"
-                md:HintAssist.Hint="Operator"
-                IsEditable="True"
-                IsEnabled="True"
-                ItemsSource="{Binding Filter.Operators}"
-                SelectedItem="{Binding Filter.PropertyOperator}"
-                StaysOpenOnEdit="True" />
-      <TextBox Margin="0,16"
-               md:HintAssist.Hint="Value"
-               Text="{Binding Filter.PropertyValue}" />
+      <TextBlock Text="{Binding Filter.Description}" TextWrapping="Wrap" />
+      <ComboBox
+        Width="Auto"
+        Margin="0,16"
+        md:HintAssist.Hint="Parameter"
+        IsEditable="True"
+        IsEnabled="True"
+        ItemsSource="{Binding Filter.Values}"
+        SelectedItem="{Binding Filter.PropertyName}"
+        StaysOpenOnEdit="True" />
+      <ComboBox
+        Width="Auto"
+        Margin="0,16"
+        md:HintAssist.Hint="Operator"
+        IsEditable="True"
+        IsEnabled="True"
+        ItemsSource="{Binding Filter.Operators}"
+        SelectedItem="{Binding Filter.PropertyOperator}"
+        StaysOpenOnEdit="True" />
+      <TextBox
+        Margin="0,16"
+        md:HintAssist.Hint="Value"
+        Text="{Binding Filter.PropertyValue}" />
     </StackPanel>
   </Grid>
 </UserControl>
