+++ Objects/Converters/ConverterRevit/ConverterRevitShared/ConversionUtils.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using DB = Autodesk.Revit.DB;
+using ElementType = Autodesk.Revit.DB.ElementType;
 using Floor = Objects.BuiltElements.Floor;
 using Level = Objects.BuiltElements.Level;
 using Parameter = Objects.BuiltElements.Revit.Parameter;
