+++ Objects/Objects/BuiltElements/Revit/ProjectInfo.cs
@@ -1,0 +1,22 @@
+﻿using Speckle.Core.Kits;
+using Speckle.Core.Models;
+using System.Collections.Generic;
+
+namespace Objects.BuiltElements.Revit
+{
+  public class ProjectInfo : Base
+  {
+    public string address { get; set; }
+    public string author { get; set; }
+    public string buildingName { get; set; }
+    public string clientName { get; set; }
+    public string issueDate { get; set; }
+    public string name { get; set; }
+    public string number { get; set; }
+    public string organizationDescription { get; set; }
+    public string organizationName { get; set; }
+    public string status { get; set; }
+
+  }
+
+}
