+++ Objects/Objects/BuiltElements/Revit/RevitElementType.cs
@@ -1,0 +1,15 @@
+﻿using Speckle.Core.Kits;
+using Speckle.Core.Models;
+using System.Collections.Generic;
+
+namespace Objects.BuiltElements.Revit
+{
+  public class RevitElementType : Base
+  {
+    public string family { get; set; }
+    public string type { get; set; }
+    public string category { get; set; }
+
+  }
+
+}
